Use Github's sort order
The existing sorting method uses a couple of heuristics to sort -
however, there are cases where the resulting sorting does not match
the order in which Github displays the files.

As a result, scrolling down the list can cause the selected file in the
tree to jump around rather than going down one-by-one.

This commit removes most of the sorting logic and instead uses the index
of the name of the file on the page, guaranteeing the tree matches
Github even if they change their ordering logic.

(cherry picked from commit ed6dbe529e64867d51896c21230e4cff6dddb855)